Tests/LibWeb/Text/expected/Editing/case-insensitive-strikethrough-command.txt

strikeThrough active?: false
strikeThrough active?: true
